Factor EmptyState to component
