+++ site/tsconfig.json
@@ -9,6 +9,6 @@
     "downlevelIteration": true,
     "esModuleInterop": true,
     "resolveJsonModule": true,
-    "isolatedModules": true,
+    "isolatedModules": true
   }
 }
